+++ BDKSwiftExampleWallet/View/ReceiveView.swift
@@ -7,21 +7,6 @@
 
 import SwiftUI
 
-class ReceiveViewModel: ObservableObject {
-    @Published var address: String = ""
-    
-    func getAddress() {
-        do {
-            let address = try BDKService.shared.getAddress()
-            self.address = address
-        } catch {
-            self.address = "Error getting address."
-        }
-    }
-    
-}
-
-
 struct ReceiveView: View {
     @ObservedObject var viewModel: ReceiveViewModel
     @State private var isCopied = false
