+++ BDKSwiftExampleWallet/View/TabHomeView.swift
@@ -39,10 +39,9 @@ struct TabHomeView: View {
                             systemImage: "arrow.up"
                         )
                     }
-                                
+                
             }
             .tint(.orange)
-            
         }
         
     }
